refactor(modules): moved module class file into src/modules directory
