+++ src/modules/auth/email.ts
@@ -1,4 +1,4 @@
-import Module from "../../module";
+import Module from "../";
 
 export interface SignUpSettingsResponse {
   /**
