+++ src/modules/auth/index.ts
@@ -1,4 +1,4 @@
-import Module from "../../module";
+import Module from "../";
 import AuthEmailModule from "./email";
 import { Client } from "../../client";
